+++ src/gob-chunk.c
@@ -48,11 +48,11 @@ int main(int argc, char *argv[])
     while ((bytes = read_bytes(STDIN_FILENO, block, BLOCK_LEN)) > 0) {
         struct hash hash;
 
-        total += bytes;
+        total += (size_t) bytes;
 
-        if (hash_state_update(&state, block, bytes) < 0)
+        if (hash_state_update(&state, block, (size_t) bytes) < 0)
             die("Unable to update hash");
-        if (store_write(&hash, &store, block, bytes) < 0)
+        if (store_write(&hash, &store, block, (size_t) bytes) < 0)
             die("Unable to store block");
         puts(hash.hex);
     }
